requirements.txt
feedparser
feedgen
openai
requests
beautifulsoup4
readability-lxml
python-dotenv
newspaper3k